# contracts/openapi/riverhog.v1.yaml
openapi: 3.1.0
info:
  title: RIVERHOG API
  version: 1.0.0
  summary: MVP API for hot/cold archival storage
  description: Authoritative archive resources are durable across service restarts unless a specific contract explicitly expires or releases them.
servers:
  - url: /v1
paths:
  /collection-uploads:
    post:
      summary: Create or resume one collection upload session
      description: Creates or resumes a collection upload session from a human-readable slug and expected logical file set. Riverhog normalizes the slug, mints the canonical collection id with either the server UTC upload timestamp or an optional caller-provided UTC timestamp, and resumes an existing non-finalized session when the normalized slug and manifest match.
      operationId: createOrResumeCollectionUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionUploadRequest"
      responses:
        "200":
          description: Collection upload session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionUploadSession"
        default:
          $ref: "#/components/responses/Error"
  /collection-upload-sessions:
    post:
      summary: Create or resume one incremental collection upload session
      description: Creates or resumes an open collection upload session from a human-readable slug without requiring a complete file manifest up front. Files are registered one at a time, uploaded through the normal collection-file upload resource, then frozen by an explicit complete call.
      operationId: createOrResumeCollectionUploadSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrResumeCollectionUploadSessionRequest"
      responses:
        "200":
          description: Collection upload session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionUploadSession"
        default:
          $ref: "#/components/responses/Error"
  /collection-upload-sessions/{collection_id}/files:
    post:
      summary: Register one file in an incremental collection upload session
      description: Adds one logical file metadata entry to an open incremental upload session. Re-registering the same path with identical metadata is idempotent; conflicting metadata is rejected. The response is scoped to the registered file; callers use the collection status endpoint for whole-session progress.
      operationId: registerCollectionUploadSessionFile
      parameters:
        - in: path
          name: collection_id
          required: true
          description: Server-minted canonical collection id returned by incremental session creation. Slash-bearing ids occupy multiple path segments.
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterCollectionUploadSessionFileRequest"
      responses:
        "200":
          description: Registered file state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionUploadSessionFileRegistration"
        default:
          $ref: "#/components/responses/Error"
  /collection-upload-sessions/{collection_id}/files/upload:
    post:
      summary: Register one file and create or resume its upload resource
      description: Adds one logical file metadata entry to an open incremental upload session, then creates or resumes that file's Riverhog-managed tus-compatible upload resource. Re-registering the same path with identical metadata is idempotent; conflicting metadata is rejected.
      operationId: createOrResumeRegisteredCollectionFileUpload
      parameters:
        - in: path
          name: collection_id
          required: true
          description: Server-minted canonical collection id returned by incremental session creation. Slash-bearing ids occupy multiple path segments.
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterCollectionUploadSessionFileRequest"
      responses:
        "200":
          description: Registered file upload session
          headers:
            Tus-Resumable:
              schema: { type: string }
            Cache-Control:
              schema: { type: string }
            Upload-Offset:
              schema: { type: integer, minimum: 0 }
            Upload-Length:
              schema: { type: integer, minimum: 0 }
            Upload-Expires:
              schema: { type: string, format: date-time }
            Location:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionUploadSessionFileUpload"
        default:
          $ref: "#/components/responses/Error"
  /collection-upload-sessions/{collection_id}/complete:
    post:
      summary: Complete one incremental collection upload session
      description: Freezes the registered file set for one open incremental upload session. Completion requires at least one registered file and all registered bytes uploaded and verified, then hands the collection to the existing archive finalization pipeline.
      operationId: completeCollectionUploadSession
      parameters:
        - in: path
          name: collection_id
          required: true
          description: Server-minted canonical collection id returned by incremental session creation. Slash-bearing ids occupy multiple path segments.
          schema: { type: string }
      responses:
        "200":
          description: Collection upload session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionUploadSession"
        default:
          $ref: "#/components/responses/Error"
  /collection-upload-sessions/{collection_id}/cancel:
    post:
      summary: Cancel one incremental collection upload session
      description: Cancels an open incremental upload session, deletes staged upload bytes, and leaves a closed audit state. Sessions cannot be canceled after the archival handoff begins.
      operationId: cancelCollectionUploadSession
      parameters:
        - in: path
          name: collection_id
          required: true
          description: Server-minted canonical collection id returned by incremental session creation. Slash-bearing ids occupy multiple path segments.
          schema: { type: string }
      responses:
        "200":
          description: Collection upload session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionUploadSession"
        default:
          $ref: "#/components/responses/Error"
  /collection-uploads/{collection_id}:
    get:
      summary: Get one collection upload session
      description: Returns the current resumable upload state for one server-minted collection id while it remains in progress. The session remains visible through `uploading`, `archiving`, and `failed`; once the collection archive package has uploaded, verified, promoted the collection, and committed the collection/archive records, this upload session is deleted and later reads return `not_found`.
      operationId: getCollectionUpload
      parameters:
        - in: path
          name: collection_id
          required: true
          description: Server-minted canonical collection id returned by collection upload creation. Slash-bearing ids occupy multiple path segments.
          schema: { type: string }
      responses:
        "200":
          description: Collection upload session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionUploadSession"
        default:
          $ref: "#/components/responses/Error"
  /collection-uploads/{collection_id}/files/{path}/upload:
    post:
      summary: Create or resume one collection-file upload resource
      description: Creates or resumes one Riverhog-managed tusd upload resource for one logical collection file and returns its direct tusd URL. Riverhog does not proxy collection upload bytes; clients send TUS requests to the returned `upload_url`.
      operationId: createOrResumeCollectionFileUpload
      parameters:
        - in: path
          name: collection_id
          required: true
          description: Server-minted canonical collection id returned by collection upload creation. Slash-bearing ids occupy multiple path segments.
          schema: { type: string }
        - in: path
          name: path
          required: true
          description: Relative logical file path within the collection.
          schema: { type: string }
      responses:
        "200":
          description: Collection-file upload session
          headers:
            Tus-Resumable:
              schema: { type: string }
            Cache-Control:
              schema: { type: string }
            Upload-Offset:
              schema: { type: integer, minimum: 0 }
            Upload-Length:
              schema: { type: integer, minimum: 0 }
            Upload-Expires:
              schema: { type: string, format: date-time }
            Location:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionFileUploadSession"
        default:
          $ref: "#/components/responses/Error"
...
  /collections:
    get:
      summary: List collection summaries
      description: Returns one page of collection summaries with optional substring and disc-redundancy filtering.
      operationId: listCollections
      parameters:
        - in: query
          name: page
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: per_page
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - in: query
          name: q
          required: false
          description: Case-insensitive substring filter over canonical collection ids.
          schema: { type: string }
        - in: query
          name: sort
          required: false
          schema: { type: string, default: id }
        - in: query
          name: order
          required: false
          schema: { type: string, default: asc }
        - in: query
          name: disc_redundancy
          required: false
          description: Exact collection disc redundancy filter.
          schema:
            type: string
            enum: [none, partial, full]
      responses:
        "200":
          description: Collection summary listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionListResponse"
        default:
          $ref: "#/components/responses/Error"
  /collections/{collection_id}:
    get:
      summary: Get one collection summary
      operationId: getCollection
      parameters:
        - in: path
          name: collection_id
          required: true
          description: Canonical collection id. Slash-bearing ids occupy multiple path segments under `/collections/`.
          schema: { type: string }
        - in: query
          name: coverage_path_limit
          required: false
          description: Maximum covered paths to return per image. Defaults to a bounded operator preview.
          schema:
            type: integer
            minimum: 0
            maximum: 100
            default: 100
      responses:
        "200":
          description: Collection summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionSummary"
        default:
          $ref: "#/components/responses/Error"
...
components:
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    FetchEntryUploadState:
      type: string
      enum: [pending, partial, byte_complete, uploaded]
...
    CoverageState:
      type: string
      enum: [none, partial, full]
    Archive:
      type: object
      required:
        [state, object_path, stored_bytes, backend, storage_class, last_uploaded_at, last_verified_at, failure]
      properties:
        state:
          type: string
          enum: [pending, uploading, uploaded, retrying, failed]
        object_path:
          oneOf:
            - type: string
            - type: "null"
        stored_bytes:
          oneOf:
            - type: integer
            - type: "null"
        backend:
          oneOf:
            - type: string
            - type: "null"
        storage_class:
          oneOf:
            - type: string
            - type: "null"
        last_uploaded_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        last_verified_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        failure:
          oneOf:
            - type: string
            - type: "null"
    CollectionManifestProofState:
      type: string
      enum: [pending, uploaded, failed]
    CollectionManifest:
      type: object
      required: [object_path, sha256, ots_object_path, ots_state]
      properties:
        object_path:
          oneOf:
            - type: string
            - type: "null"
        sha256:
          oneOf:
            - type: string
              pattern: "^[0-9a-f]{64}$"
            - type: "null"
        ots_object_path:
          oneOf:
            - type: string
            - type: "null"
        ots_state:
          $ref: "#/components/schemas/CollectionManifestProofState"
...
    CollectionSummary:
      type: object
      required:
        [
          id,
          files,
          bytes,
          hot_bytes,
          archive,
          collection_manifest,
          archive_format,
          compression,
          disc_coverage,
          disc_redundancy,
          image_coverage,
        ]
      properties:
        id:
          type: string
          description: Canonical collection id. May contain `/` and is prefix-free with every other collection id.
        files: { type: integer }
        bytes: { type: integer }
        hot_bytes: { type: integer }
        archive:
          $ref: "#/components/schemas/Archive"
        collection_manifest:
          $ref: "#/components/schemas/CollectionManifest"
        archive_format:
          type: string
        compression:
          type: string
        disc_coverage:
          $ref: "#/components/schemas/Coverage"
        disc_redundancy:
          $ref: "#/components/schemas/Coverage"
        image_coverage:
          type: array
          items:
            $ref: "#/components/schemas/CollectionCoverageImage"
    CollectionListItem:
      type: object
      required:
        [
          id,
          files,
          bytes,
          hot_bytes,
          archive,
          collection_manifest,
          archive_format,
          compression,
          disc_coverage,
          disc_redundancy,
        ]
      properties:
        id:
          type: string
          description: Canonical collection id. May contain `/` and is prefix-free with every other collection id.
        files: { type: integer }
        bytes: { type: integer }
        hot_bytes: { type: integer }
        archive:
          $ref: "#/components/schemas/Archive"
        collection_manifest:
          $ref: "#/components/schemas/CollectionManifest"
        archive_format:
          type: string
        compression:
          type: string
        disc_coverage:
          $ref: "#/components/schemas/Coverage"
        disc_redundancy:
          $ref: "#/components/schemas/Coverage"
    Coverage:
      type: object
      required: [state, bytes]
      properties:
        state:
          $ref: "#/components/schemas/CoverageState"
        bytes:
          type: integer
          minimum: 0
    CollectionListResponse:
      type: object
      required: [page, per_page, total, pages, collections]
      properties:
        page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        pages:
          type: integer
          minimum: 0
        collections:
          type: array
          items:
            $ref: "#/components/schemas/CollectionListItem"
    CollectionUploadFile:
      type: object
      required: [path, bytes, sha256]
      properties:
        path:
          type: string
          description: Relative logical file path within the collection source tree.
        bytes: { type: integer }
        sha256:
          type: string
          description: Logical-file SHA-256 digest for final verification.
    CollectionUploadNotify:
      type: object
      properties:
        enabled:
          type: boolean
          description: Whether Riverhog should deliver collection lifecycle notifications for this upload.
        recipients:
          type: array
          items:
            type: string
          description: Deployment-owned recipient ids resolved through Riverhog notification config.
    CollectionUploadRequest:
      type: object
      required: [slug, files]
      properties:
        slug:
          type: string
          description: Human-readable collection slug. Riverhog normalizes this and mints the canonical collection id.
        upload_timestamp:
          oneOf:
            - type: string
            - type: "null"
          description: Optional UTC basic timestamp in `YYYYMMDDTHHMMSSZ` form to use in the canonical collection id for migration of older archival sets. Slug remains required.
        ingest_source:
          oneOf:
            - type: string
            - type: "null"
          description: Descriptive source hint from the client. Never part of collection identity.
        notify:
          oneOf:
            - $ref: "#/components/schemas/CollectionUploadNotify"
            - type: "null"
          description: Optional collection lifecycle notification routing metadata.
        files:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CollectionUploadFile"
    CreateOrResumeCollectionUploadSessionRequest:
      type: object
      required: [slug]
      properties:
        slug:
          type: string
          description: Human-readable collection slug. Riverhog normalizes this and mints or resumes the canonical collection id.
        upload_timestamp:
          oneOf:
            - type: string
            - type: "null"
          description: Optional UTC basic timestamp in `YYYYMMDDTHHMMSSZ` form to use in the canonical collection id for migration of older archival sets. Slug remains required.
        ingest_source:
          oneOf:
            - type: string
            - type: "null"
          description: Descriptive source hint from the client. Never part of collection identity.
        notify:
          oneOf:
            - $ref: "#/components/schemas/CollectionUploadNotify"
            - type: "null"
          description: Optional collection lifecycle notification routing metadata.
    RegisterCollectionUploadSessionFileRequest:
      allOf:
        - $ref: "#/components/schemas/CollectionUploadFile"
    CollectionUploadFileState:
      type: object
      required: [path, bytes, sha256, upload_state, uploaded_bytes, upload_state_expires_at]
      properties:
        path: { type: string }
        bytes: { type: integer }
        sha256: { type: string }
        upload_state:
          $ref: "#/components/schemas/FetchEntryUploadState"
        uploaded_bytes:
          type: integer
          description: Accepted upload bytes for this logical file's plaintext stream.
        upload_state_expires_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
    CollectionUploadSessionFileRegistration:
      type: object
      required: [collection_id, ingest_source, state, file]
      properties:
        collection_id:
          type: string
          description: Server-minted canonical collection id for this upload session.
        ingest_source:
          oneOf:
            - type: string
            - type: "null"
        state:
          type: string
          enum: [open, uploading]
        file:
          $ref: "#/components/schemas/CollectionUploadFileState"
    CollectionUploadSessionFileUpload:
      allOf:
        - $ref: "#/components/schemas/CollectionUploadSessionFileRegistration"
        - $ref: "#/components/schemas/CollectionFileUploadSession"
    CollectionUploadSession:
      type: object
      required:
        [
          collection_id,
          ingest_source,
          state,
          files_total,
          files_pending,
          files_partial,
          files_uploaded,
          hot_promoted_files,
          bytes_total,
          uploaded_bytes,
          hot_promoted_bytes,
          missing_bytes,
          upload_state_expires_at,
          files,
          collection,
        ]
      properties:
        collection_id:
          type: string
          description: Server-minted canonical collection id for this upload session.
        ingest_source:
          oneOf:
            - type: string
            - type: "null"
        state:
          type: string
          enum: [open, uploading, archiving, finalized, failed, canceled, expired]
        files_total: { type: integer }
        files_pending: { type: integer }
        files_partial: { type: integer }
        files_uploaded: { type: integer }
        hot_promoted_files:
          type: integer
          description: Files verified in the hot collection namespace during finalization.
        bytes_total: { type: integer }
        uploaded_bytes: { type: integer }
        hot_promoted_bytes:
          type: integer
          description: Bytes verified in the hot collection namespace during finalization.
        missing_bytes: { type: integer }
        upload_state_expires_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        latest_failure:
          oneOf:
            - type: string
            - type: "null"
          description: Latest collection archive failure, when archiving exhausted retries.
        archive_phase:
          oneOf:
            - type: string
            - type: "null"
          description: Current long-running archive phase for upload finalization.
        archive_phase_updated_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        archive_object_path:
          oneOf:
            - type: string
            - type: "null"
        archive_uploaded_bytes:
          oneOf:
            - type: integer
            - type: "null"
        archive_total_bytes:
          oneOf:
            - type: integer
            - type: "null"
        archive_uploaded_parts:
          oneOf:
            - type: integer
            - type: "null"
        archive_total_parts:
          oneOf:
            - type: integer
            - type: "null"
        notify:
          oneOf:
            - $ref: "#/components/schemas/CollectionUploadNotify"
            - type: "null"
        files:
          type: array
          items:
            $ref: "#/components/schemas/CollectionUploadFileState"
        collection:
          oneOf:
            - $ref: "#/components/schemas/CollectionSummary"
            - type: "null"
    CollectionFileUploadSession:
      type: object
      required: [path, protocol, upload_url, offset, length, checksum_algorithm, expires_at]
      properties:
        path: { type: string }
        protocol:
          type: string
          const: tus
        upload_url:
          type: string
          format: uri
        offset:
          type: integer
          description: Current accepted byte offset in the logical file's upload stream.
        length:
          type: integer
          description: Total logical-file byte length for the upload stream.
        checksum_algorithm:
          type: string
          const: sha256
        expires_at:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
...
    CollectionCoverageImage:
      type: object
      required:
        [
          id,
          filename,
          disc_redundancy_state,
          discs_required,
          discs_registered,
          discs_verified,
          discs_missing,
          covered_paths,
          discs,
        ]
      properties:
        id:
          type: string
          pattern: "^[0-9]{8}T[0-9]{6}Z$"
        filename:
          type: string
        disc_redundancy_state:
          $ref: "#/components/schemas/CoverageState"
        discs_required:
          type: integer
          minimum: 1
        discs_registered:
          type: integer
          minimum: 0
        discs_verified:
          type: integer
          minimum: 0
        discs_missing:
          type: integer
          minimum: 0
        covered_paths:
          type: array
          items:
            type: string
          description: Relative collection paths within this collection that the finalized image represents.
        covered_paths_total:
          type: integer
          minimum: 0
          description: Total covered path count before any `coverage_path_limit` truncation.
        discs:
          type: array
          items:
            $ref: "#/components/schemas/DiscSummary"
    DiscSummary:
      type: object
      required:
        [
          disc_id,
          label_text,
          location,
          created_at,
          state,
          verification_state,
          history,
          image_id,
        ]
      properties:
        disc_id:
          type: string
          description: Generated disc id, unique within one finalized image.
        label_text:
          type: string
          description: Exact operator-facing disc label text for this generated disc slot.
        location:
          type: string
          nullable: true
          description: Mutable physical location label. Not part of disc identity.
        created_at: { type: string, format: date-time }
        state:
          type: string
          enum: [needed, burning, verified, registered, lost, damaged, retired]
        verification_state:
          type: string
          enum: [pending, verified, failed]
        history:
          type: array
          items:
            $ref: "#/components/schemas/DiscHistory"
        image_id:
          type: string
          pattern: "^[0-9]{8}T[0-9]{6}Z$"
          description: Finalized image id for this disc.
        filename:
          type: string
          nullable: true
          description: Stable ISO filename for the finalized image artifact.
...
    DiscHistory:
      type: object
      required: [at, event, state, verification_state, location]
      properties:
        at: { type: string, format: date-time }
        event: { type: string }
        state:
          type: string
          enum: [needed, burning, verified, registered, lost, damaged, retired]
        verification_state:
          type: string
          enum: [pending, verified, failed]
        location:
          type: string
          nullable: true
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }